+++ objectified-ui/public/WHATS_NEW.md
@@ -14,6 +14,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Added background configurations to the canvas
   - Added ability to enable/disable grid
   - Added ability to search using regular expressions in the canvas
+  - Nodes now show property references even when not expanded
   - Improved search to include properties and descriptions in the results
   - Canvas changes also apply to the Paths canvas
 - Paths Improvements: